requirements.txt
nbformat==5.10.4
prophet==1.1.6
pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.2
yfinance==0.2.49
plotly==5.24.1
flask==3.0.3
datetime==5.5
sqlalchemy==2.0.36
pymysql==1.1.1 
psycopg2-binary==2.9.10
requests==2.32.3 
spotipy==2.24.0 
ipyleaflet==0.19.2 
ipywidgets==8.1.5 
simpy==4.1.1 
textblob==0.18.0.post0 
vadersentiment==3.3.2 
regex==2024.11.6 
newsapi==0.1.1 
langdetect==1.0.9 
newsapi-python==0.2.7
beautifulsoup4==4.12.3
PyMuPDF==1.24.13
PyPDF2==3.0.1
openai==0.28
tiktoken==0.8.0
python-dotenv==1.0.1
scikit-learn==1.5.2
requests==2.32.3
pdfkit==1.0.0
transformers==4.46.2
fpdf==1.7.2
feedparser==6.0.11
